src/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
influxdb==5.2.3
paho-mqtt==1.5.0
pendulum==2.1.0
python-dateutil==2.8.1
pytz==2019.3
pytzdata==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
